+++ programs/dynamic-bonding-curve/src/error.rs
@@ -112,4 +112,19 @@ pub enum PoolError {
 
     #[msg("Invalid creator trading fee percentage")]
     InvalidCreatorTradingFeePercentage,
+
+    #[msg("Invalid new creator")]
+    InvalidNewCreator,
+
+    #[msg("Invalid token update authority option")]
+    InvalidTokenUpdateAuthorityOption,
+
+    #[msg("Invalid account for the instruction")]
+    InvalidAccount,
+
+    #[msg("Invalid migrator fee percentage")]
+    InvalidMigratorFeePercentage,
+
+    #[msg("Migration fee has been withdraw")]
+    MigrationFeeHasBeenWithdraw,
 }