+++ programs/dynamic-bonding-curve/src/event.rs
@@ -149,3 +149,23 @@ pub struct EvtWithdrawLeftover {
     pub leftover_receiver: Pubkey,
     pub leftover_amount: u64,
 }
+
+#[event]
+pub struct EvtUpdatePoolCreator {
+    pub pool: Pubkey,
+    pub creator: Pubkey,
+    pub new_creator: Pubkey,
+}
+
+#[event]
+pub struct EvtWithdrawMigrationFee {
+    pub pool: Pubkey,
+    pub fee: u64,
+    pub flag: u8,
+}
+
+#[event]
+pub struct EvtPartnerWithdrawMigrationFee {
+    pub pool: Pubkey,
+    pub fee: u64,
+}