+++ programs/dynamic-bonding-curve/src/instructions/creator/mod.rs
@@ -4,3 +4,5 @@ pub mod ix_claim_creator_trading_fee;
 pub use ix_claim_creator_trading_fee::*;
 pub mod ix_withdraw_creator_surplus;
 pub use ix_withdraw_creator_surplus::*;
+pub mod ix_transfer_pool_creator;
+pub use ix_transfer_pool_creator::*;