+++ programs/dynamic-bonding-curve/src/instructions/migration/meteora_damm/meteora_damm_claim_lp_token.rs
@@ -84,7 +84,7 @@ pub fn handle_migrate_meteora_damm_claim_lp_token<'info>(
     let mut migration_metadata = ctx.accounts.migration_metadata.load_mut()?;
 
     let is_partner = ctx.accounts.owner.key() == migration_metadata.partner;
-    let is_creator = ctx.accounts.owner.key() == migration_metadata.pool_creator;
+    let is_creator = ctx.accounts.owner.key() == virtual_pool.creator;
 
     let lp_to_claim = match (is_partner, is_creator) {
         (true, true) => migration_metadata.claim_as_self_partnered_creator()?,