+++ tests/claim_and_lock_lp_on_meteora_damm.tests.ts
@@ -103,6 +103,11 @@ async function createPartnerConfig(
     migrationFeeOption: 0,
     tokenSupply: null,
     creatorTradingFeePercentage: 0,
+    tokenUpdateAuthority: 0,
+    migrationFee: {
+      feePercentage: 0,
+      creatorFeePercentage: 0,
+    },
     padding0: [],
     padding: [],
     curve: curves,