+++ tests/create_pool_with_token2022.tests.ts
@@ -119,6 +119,11 @@ describe("Create pool with token2022", () => {
             migrationFeeOption: 0,
             tokenSupply: null,
             creatorTradingFeePercentage: 0,
+            tokenUpdateAuthority: 0,
+            migrationFee: {
+                feePercentage: 0,
+                creatorFeePercentage: 0,
+            },
             padding0: [],
             padding: [],
             curve: curves,