+++ tests/fee_swap.tests.ts
@@ -101,6 +101,11 @@ describe("Fee Swap test", () => {
         migrationFeeOption: 0,
         tokenSupply: null,
         creatorTradingFeePercentage: 0,
+        tokenUpdateAuthority: 0,
+        migrationFee: {
+          feePercentage: 0,
+          creatorFeePercentage: 0,
+        },
         padding0: [],
         padding: [],
         curve: curves,
@@ -446,6 +451,11 @@ describe("Fee Swap test", () => {
         migrationFeeOption: 0,
         tokenSupply: null,
         creatorTradingFeePercentage: 0,
+        tokenUpdateAuthority: 0,
+        migrationFee: {
+          feePercentage: 0,
+          creatorFeePercentage: 0,
+        },
         padding0: [],
         padding: [],
         curve: curves,