+++ tests/simulate_cu_swap.tests.ts
@@ -80,6 +80,11 @@ describe("Simulate CU swap", () => {
         migrationFeeOption: 0,
         tokenSupply: null,
         creatorTradingFeePercentage: 0,
+        tokenUpdateAuthority: 0,
+        migrationFee: {
+          feePercentage: 0,
+          creatorFeePercentage: 0,
+        },
         padding0: [],
         padding: [],
         curve: curves,